Remove some dependencies on BaseStudyController and SpecimenUtils. Move a bunch of classes from study-api to specimen-main.
