+++ specimen/src/org/labkey/specimen/actions/ManageReqsBean.java
@@ -1,4 +1,4 @@
-package org.labkey.api.specimen.actions;
+package org.labkey.specimen.actions;
 
 import org.labkey.api.data.Container;
 import org.labkey.api.security.User;
