+++ specimen/src/org/labkey/specimen/actions/SpecimenController2.java
@@ -32,24 +32,22 @@ import org.labkey.api.security.permissions.AdminPermission;
 import org.labkey.api.security.permissions.ReadPermission;
 import org.labkey.api.specimen.SpecimenManagerNew;
 import org.labkey.api.specimen.SpecimenRequestManager;
-import org.labkey.api.specimen.SpecimenSearchWebPart;
+import org.labkey.specimen.view.SpecimenSearchWebPart;
 import org.labkey.api.specimen.Vial;
 import org.labkey.api.specimen.actions.IdForm;
-import org.labkey.api.specimen.actions.ManageReqsBean;
 import org.labkey.api.specimen.actions.SpecimenEventBean;
-import org.labkey.api.specimen.actions.ViewRequestsHeaderBean;
 import org.labkey.api.specimen.importer.RequestabilityManager;
 import org.labkey.api.specimen.importer.SimpleSpecimenImporter;
 import org.labkey.api.specimen.model.SpecimenRequestEvent;
-import org.labkey.api.specimen.pipeline.SpecimenArchive;
-import org.labkey.api.specimen.query.SpecimenEventQueryView;
-import org.labkey.api.specimen.query.SpecimenRequestQueryView;
+import org.labkey.specimen.pipeline.SpecimenArchive;
+import org.labkey.specimen.query.SpecimenEventQueryView;
+import org.labkey.specimen.query.SpecimenRequestQueryView;
 import org.labkey.api.specimen.requirements.SpecimenRequestRequirement;
 import org.labkey.api.specimen.requirements.SpecimenRequestRequirementProvider;
 import org.labkey.api.specimen.requirements.SpecimenRequestRequirementType;
-import org.labkey.api.specimen.security.permissions.ManageDisplaySettingsPermission;
-import org.labkey.api.specimen.security.permissions.ManageNotificationsPermission;
-import org.labkey.api.specimen.security.permissions.ManageRequestRequirementsPermission;
+import org.labkey.specimen.security.permissions.ManageDisplaySettingsPermission;
+import org.labkey.specimen.security.permissions.ManageNotificationsPermission;
+import org.labkey.specimen.security.permissions.ManageRequestRequirementsPermission;
 import org.labkey.api.specimen.security.permissions.ManageRequestSettingsPermission;
 import org.labkey.api.specimen.security.permissions.RequestSpecimensPermission;
 import org.labkey.api.specimen.settings.DisplaySettings;
