+++ specimen/src/org/labkey/specimen/importer/SpecimenSettingsImporter.java
@@ -23,7 +23,7 @@ import org.labkey.api.security.SecurityManager;
 import org.labkey.api.specimen.SpecimenRequestManager;
 import org.labkey.api.specimen.SpecimenRequestManager.SpecimenRequestInput;
 import org.labkey.api.specimen.SpecimenRequestStatus;
-import org.labkey.api.specimen.actions.ManageReqsBean;
+import org.labkey.specimen.actions.ManageReqsBean;
 import org.labkey.api.specimen.importer.RequestabilityManager;
 import org.labkey.api.specimen.location.LocationCache;
 import org.labkey.api.specimen.location.LocationImpl;
