+++ specimen/src/org/labkey/specimen/security/roles/SpecimenCoordinatorRole.java
@@ -19,11 +19,11 @@ import org.labkey.api.security.Group;
 import org.labkey.api.security.SecurityManager;
 import org.labkey.api.security.permissions.EditSharedViewPermission;
 import org.labkey.api.specimen.security.permissions.EditSpecimenDataPermission;
-import org.labkey.api.specimen.security.permissions.LockSpecimensPermission;
-import org.labkey.api.specimen.security.permissions.ManageDisplaySettingsPermission;
+import org.labkey.specimen.security.permissions.LockSpecimensPermission;
+import org.labkey.specimen.security.permissions.ManageDisplaySettingsPermission;
 import org.labkey.api.specimen.security.permissions.ManageNewRequestFormPermission;
-import org.labkey.api.specimen.security.permissions.ManageNotificationsPermission;
-import org.labkey.api.specimen.security.permissions.ManageRequestRequirementsPermission;
+import org.labkey.specimen.security.permissions.ManageNotificationsPermission;
+import org.labkey.specimen.security.permissions.ManageRequestRequirementsPermission;
 import org.labkey.api.specimen.security.permissions.ManageRequestSettingsPermission;
 import org.labkey.api.specimen.security.permissions.ManageRequestStatusesPermission;
 import org.labkey.api.specimen.security.permissions.ManageRequestsPermission;
