+++ specimen/src/org/labkey/specimen/view/manageDefaultReqs.jsp
@@ -15,14 +15,14 @@
  * limitations under the License.
  */
 %>
-<%@ page import="org.labkey.api.specimen.actions.ManageReqsBean" %>
 <%@ page import="org.labkey.api.specimen.model.SpecimenRequestActor" %>
 <%@ page import="org.labkey.api.specimen.requirements.SpecimenRequestRequirement" %>
-<%@ page import="org.labkey.api.study.StudyUrls"%>
+<%@ page import="org.labkey.api.study.StudyUrls" %>
 <%@ page import="org.labkey.api.util.HtmlString"%>
-<%@ page import="org.labkey.api.view.ActionURL" %>
+<%@ page import="org.labkey.api.view.ActionURL"%>
 <%@ page import="org.labkey.api.view.HttpView" %>
 <%@ page import="org.labkey.api.view.JspView" %>
+<%@ page import="org.labkey.specimen.actions.ManageReqsBean" %>
 <%@ page import="org.labkey.specimen.actions.SpecimenController2" %>
 <%@ page import="org.labkey.specimen.actions.SpecimenController2.ManageDefaultReqsAction" %>
 <%@ taglib prefix="labkey" uri="http://www.labkey.org/taglib" %>
