+++ specimen/src/org/labkey/specimen/view/viewRequestsHeader.jsp
@@ -16,11 +16,11 @@
  */
 %>
 <%@ page import="org.labkey.api.specimen.SpecimenRequestStatus" %>
-<%@ page import="org.labkey.api.specimen.actions.ViewRequestsHeaderBean" %>
 <%@ page import="org.labkey.api.view.ActionURL" %>
 <%@ page import="org.labkey.api.view.HttpView" %>
 <%@ page import="org.labkey.api.view.JspView" %>
 <%@ page import="org.labkey.api.view.ViewContext" %>
+<%@ page import="org.labkey.specimen.actions.ViewRequestsHeaderBean" %>
 <%@ page extends="org.labkey.api.jsp.JspBase" %>
 <%
     JspView<ViewRequestsHeaderBean> me = (JspView<ViewRequestsHeaderBean>) HttpView.currentView();
