+++ study/src/org/labkey/study/controllers/specimen/SpecimenUtils.java
@@ -18,7 +18,6 @@ package org.labkey.study.controllers.specimen;
 
 import org.apache.commons.lang3.StringUtils;
 import org.labkey.api.action.SpringActionController;
-import org.labkey.api.annotations.Migrate;
 import org.labkey.api.attachments.Attachment;
 import org.labkey.api.attachments.AttachmentService;
 import org.labkey.api.data.ActionButton;
@@ -28,7 +27,6 @@ import org.labkey.api.data.CompareType;
 import org.labkey.api.data.Container;
 import org.labkey.api.data.DataColumn;
 import org.labkey.api.data.DataRegion;
-import org.labkey.api.data.DataRegionSelection;
 import org.labkey.api.data.DisplayColumn;
 import org.labkey.api.data.ExcelWriter;
 import org.labkey.api.data.MenuButton;
@@ -45,10 +43,7 @@ import org.labkey.api.query.QueryDefinition;
 import org.labkey.api.query.QueryService;
 import org.labkey.api.security.User;
 import org.labkey.api.security.permissions.AdminPermission;
-import org.labkey.api.specimen.AmbiguousLocationException;
 import org.labkey.api.specimen.RequestEventType;
-import org.labkey.api.specimen.RequestedSpecimens;
-import org.labkey.api.specimen.SpecimenManagerNew;
 import org.labkey.api.specimen.SpecimenMigrationService;
 import org.labkey.api.specimen.SpecimenRequestManager;
 import org.labkey.api.specimen.SpecimenSchema;
@@ -88,7 +83,6 @@ import org.labkey.api.view.GridView;
 import org.labkey.api.view.NavTree;
 import org.labkey.api.view.RedirectException;
 import org.labkey.api.view.ViewContext;
-import org.labkey.study.controllers.BaseStudyController;
 import org.labkey.study.controllers.StudyController;
 import org.labkey.study.model.DatasetDefinition;
 import org.labkey.study.model.ParticipantGroupManager;
@@ -106,14 +100,11 @@ import java.io.IOException;
 import java.io.Writer;
 import java.sql.SQLException;
 import java.util.ArrayList;
-import java.util.Arrays;
 import java.util.Collection;
 import java.util.Collections;
 import java.util.HashMap;
-import java.util.HashSet;
 import java.util.List;
 import java.util.Map;
-import java.util.Set;
 
 /**
  * User: brittp
@@ -531,90 +522,6 @@ public class SpecimenUtils
     {
         if (!SettingsManager.get().isSpecimenRequestEnabled(getContainer(), checkExistingStatuses))
             throw new RedirectException(SpecimenMigrationService.get().getSpecimenRequestConfigRequiredURL(getContainer()));
-    }
-
-
-    public List<Vial> getSpecimensFromRowIds(long[] requestedSampleIds)
-    {
-        List<Vial> requestedVials = null;
-
-        if (requestedSampleIds != null)
-        {
-            List<Vial> vials = new ArrayList<>();
-            for (long requestedSampleId : requestedSampleIds)
-            {
-                Vial current = SpecimenManagerNew.get().getVial(getContainer(), getUser(), requestedSampleId);
-                if (current != null)
-                    vials.add(current);
-            }
-            requestedVials = vials;
-        }
-
-        return requestedVials;
-    }
-
-    public List<Vial> getSpecimensFromGlobalUniqueIds(Set<String> globalUniqueIds)
-    {
-        User user = getUser();
-        Container container = getContainer();
-        List<Vial> requestedVials = null;
-
-        if (globalUniqueIds != null)
-        {
-            List<Vial> vials = new ArrayList<>();
-            for (String globalUniqueId : globalUniqueIds)
-            {
-                Vial match = SpecimenManagerNew.get().getVial(container, user, globalUniqueId);
-                if (match != null)
-                    vials.add(match);
-            }
-            requestedVials = new ArrayList<>(vials);
-        }
-
-        return requestedVials;
-    }
-
-    public List<Vial> getSpecimensFromRowIds(Collection<String> ids)
-    {
-        return getSpecimensFromRowIds(BaseStudyController.toLongArray(ids));
-    }
-
-    public List<Vial> getSpecimensFromPost(boolean fromGroupedView, boolean onlyAvailable)
-    {
-        Set<String> formValues = null;
-        if ("POST".equalsIgnoreCase(getViewContext().getRequest().getMethod()))
-            formValues = DataRegionSelection.getSelected(getViewContext(), true);
-
-        if (formValues == null || formValues.isEmpty())
-            return null;
-
-        List<Vial> selectedVials;
-        if (fromGroupedView)
-        {
-            Map<String, List<Vial>> keyToVialMap =
-                    SpecimenManagerNew.get().getVialsForSpecimenHashes(getContainer(), getUser(),  formValues, onlyAvailable);
-            List<Vial> vials = new ArrayList<>();
-            for (List<Vial> vialList : keyToVialMap.values())
-                vials.addAll(vialList);
-            selectedVials = new ArrayList<>(vials);
-        }
-        else
-            selectedVials = getSpecimensFromRowIds(formValues);
-        return selectedVials;
-    }
-
-    public RequestedSpecimens getRequestableByVialRowIds(Set<String> rowIds)
-    {
-        Set<Long> ids = new HashSet<>();
-        Arrays.stream(BaseStudyController.toLongArray(rowIds)).forEach(ids::add);
-        List<Vial> requestedSpecimens = SpecimenManagerNew.get().getRequestableVials(getContainer(), getUser(), ids);
-        return new RequestedSpecimens(requestedSpecimens);
-    }
-
-    public RequestedSpecimens getRequestableByVialGlobalUniqueIds(Set<String> globalUniqueIds)
-    {
-        List<Vial> requestedSpecimens = getSpecimensFromGlobalUniqueIds(globalUniqueIds);
-        return new RequestedSpecimens(requestedSpecimens);
     }
 
     public GridView getRequestEventGridView(HttpServletRequest request, BindException errors, SimpleFilter filter)
